Update NuGet versions, refactor Amount type to decimal
Updated NuGet package versions in project files:
- `Microsoft.Extensions.DependencyInjection.Abstractions` and `Microsoft.Extensions.Http` updated to `9.0.6` in `Traffic.APIClient.csproj`.
- `Microsoft.AspNetCore.OpenApi` updated to `9.0.6` in `Traffic.API.csproj`.

Refactored `Amount` property type from `double` to `decimal`:
- Changes applied in `TrafficDto.cs` across `Traffic.APIClient` and `Traffic.API.Data` namespaces.
- Updated parsing logic in `TrafficProxy.cs` to use `decimal.Parse`.

Adjusted `Trend` property in `TrafficDto.cs` (`Traffic.API.Data`):
- Removed default value of `string.Empty`, leaving it uninitialized.

+++ src/api/Data/TrafficDto.cs
@@ -2,7 +2,7 @@ namespace Traffic.API.Data
 {
     public class TrafficDto
     {
-        public double Amount { get; set; } = -1;
+        public decimal Amount { get; set; } = -1;
         public string Unit { get; set; } = "Kilometers";
         public DateTime TimestampUpdated { get; set; } = DateTime.Parse("1900-01-01T01:00:00");
         public string Source { get; set; } = "verkeerscentrum.be";
